refactor(metrics): metrics をマークダウンに書き込む処理を generateTsg 近くへ移動
